Split SettingsScreen into two
SettingsPage: internal, SettingsScreen: specifying the settings
