+++ Main/include/GameConfig.hpp
@@ -1,7 +1,6 @@
 #pragma once
 #include "Shared/Config.hpp"
 #include "Input.hpp"
-#include <unordered_set>
 
 DefineEnum(GameConfigKeys,
 		   // Version of the config
